+++ apps/common-app/runtime-tests/ReJest/RuntimeTestsApi.ts
@@ -209,8 +209,8 @@ export async function waitForNotifications(
   return notificationRegistry.waitForNotifications(names, timeout);
 }
 
-export function getWorkletRuntimeFromPool(name: string) {
-  return workletRuntimePool.getOrCreateWorkletRuntime(name);
+export function getWorkletRuntimesFromPool(count: number) {
+  return workletRuntimePool.getOrCreateWorkletRuntimes(count);
 }
 
 export function expect(value: TestValue) {